+++ sdk/ai/ai-projects/src/api/index.ts
@@ -5,8 +5,15 @@ export { createAIProject, AIProjectContext } from "./aiProjectContext.js";
 
 export { AIProjectClientOptionalParams };
 
-export interface DatasetUploadOptionalParams {
-  // Optional connection name for the storage account to be used
+/** Options for uploading datasets */
+export interface DatasetUploadOptionalOptions {
+  /** Optional connection name for the storage account to be used */
   connectionName?: string;
+  /** Only for folder uploads. A regex pattern to filter files to be uploaded. Only files matching the pattern will be uploaded */
+  filePattern?: RegExp;
+}
+
+/** Options for dataset upload operations that are internal and may not be exposed to all users */
+export interface DatasetUploadInternalOptionalParams extends DatasetUploadOptionalOptions {
   projectOptions?: AIProjectClientOptionalParams;
 }